+++ src/user/user_model.cc
@@ -622,7 +622,7 @@ void mjCModel::RemoveFromTree(const mjCFrame& frame) {
 
   // delete the plugins created by the removed elements
   for (mjCBody* child : bodies) { DeleteSubtreePlugin(child); }
-  for (mjCGeom* geom : geoms) { DeleteImplicitPlugin(geom->plugin); }
+  for (mjCGeom* geom : geoms) { DeleteImplicitPlugin(geom->spec.plugin); }
 }
 
 
@@ -822,7 +822,7 @@ void mjCModel::operator-=(mjsElement* el) {
 
     case mjOBJ_GEOM: {
       mjCGeom* geom = static_cast<mjCGeom*>(el);
-      DeleteImplicitPlugin(geom->plugin);
+      DeleteImplicitPlugin(geom->spec.plugin);
       deletefromlist(&(geom->body->geoms), el);
       break;
     }
@@ -845,21 +845,21 @@ void mjCModel::operator-=(mjsElement* el) {
 
     case mjOBJ_MESH: {
       mjCMesh* mesh = static_cast<mjCMesh*>(el);
-      DeleteImplicitPlugin(mesh->plugin);
+      DeleteImplicitPlugin(mesh->spec.plugin);
       deletefromlist(object_lists_[mjOBJ_MESH], el);
       break;
     }
 
     case mjOBJ_ACTUATOR: {
       mjCActuator* actuator = static_cast<mjCActuator*>(el);
-      DeleteImplicitPlugin(actuator->plugin);
+      DeleteImplicitPlugin(actuator->spec.plugin);
       deletefromlist(object_lists_[mjOBJ_ACTUATOR], el);
       break;
     }
 
     case mjOBJ_SENSOR: {
       mjCSensor* sensor = static_cast<mjCSensor*>(el);
-      DeleteImplicitPlugin(sensor->plugin);
+      DeleteImplicitPlugin(sensor->spec.plugin);
       deletefromlist(object_lists_[mjOBJ_SENSOR], el);
       break;
     }